# docs/openapi.swagger.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "AI Audit System API",
                 "version":  "2.3.0",
                 "description":  "REST API for AI-powered quality audit system with GPT-5",
                 "contact":  {
                                 "name":  "API Support"
                             }
             },
    "servers":  [
                    {
                        "url":  "http://localhost:3002",
                        "description":  "Development server"
                    }
                ],
...
    "paths":  {
...
                  "/api/audits/{audit_id}/chat/history":  {
                                                              "get":  {
                                                                          "tags":  [
                                                                                       "Chat"
                                                                                   ],
                                                                          "summary":  "Get chat conversation history for an audit",
                                                                          "parameters":  [
                                                                                             {
                                                                                                 "in":  "path",
                                                                                                 "name":  "audit_id",
                                                                                                 "required":  true,
                                                                                                 "schema":  {
                                                                                                                "type":  "string"
                                                                                                            }
                                                                                             }
                                                                                         ],
                                                                          "responses":  {
                                                                                            "200":  {
                                                                                                        "description":  "Chat conversation history"
                                                                                                    },
                                                                                            "404":  {
                                                                                                        "description":  "Audit not found"
                                                                                                    }
                                                                                        }
                                                                      }
                                                          },
                  "/api/audits/{audit_id}/chat":  {
                                                      "post":  {
                                                                   "tags":  [
                                                                                "Chat"
                                                                            ],
                                                                   "summary":  "Chat about a specific audit (streaming)",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "in":  "path",
                                                                                          "name":  "audit_id",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "requestBody":  {
                                                                                       "required":  true,
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "type":  "object",
                                                                                                                                               "required":  [
                                                                                                                                                                "message"
                                                                                                                                                            ],
                                                                                                                                               "properties":  {
                                                                                                                                                                  "message":  {
                                                                                                                                                                                  "type":  "string"
                                                                                                                                                                              }
                                                                                                                                                              }
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "SSE stream of AI response",
                                                                                                 "content":  {
                                                                                                                 "text/event-stream":  {

                                                                                                                                       }
                                                                                                             }
                                                                                             }
                                                                                 }
                                                               }
                                                  },
                  "/api/fiches/{fiche_id}/chat/history":  {
                                                              "get":  {
                                                                          "tags":  [
                                                                                       "Chat"
                                                                                   ],
                                                                          "summary":  "Get chat conversation history for a fiche",
                                                                          "parameters":  [
                                                                                             {
                                                                                                 "in":  "path",
                                                                                                 "name":  "fiche_id",
                                                                                                 "required":  true,
                                                                                                 "schema":  {
                                                                                                                "type":  "string"
                                                                                                            }
                                                                                             }
                                                                                         ],
                                                                          "responses":  {
                                                                                            "200":  {
                                                                                                        "description":  "Chat conversation history"
                                                                                                    },
                                                                                            "404":  {
                                                                                                        "description":  "No conversation found"
                                                                                                    }
                                                                                        }
                                                                      }
                                                          },
                  "/api/fiches/{fiche_id}/chat":  {
                                                      "post":  {
                                                                   "tags":  [
                                                                                "Chat"
                                                                            ],
                                                                   "summary":  "Chat about a fiche and all its audits (streaming)",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "in":  "path",
                                                                                          "name":  "fiche_id",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "requestBody":  {
                                                                                       "required":  true,
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "type":  "object",
                                                                                                                                               "required":  [
                                                                                                                                                                "message"
                                                                                                                                                            ],
                                                                                                                                               "properties":  {
                                                                                                                                                                  "message":  {
                                                                                                                                                                                  "type":  "string"
                                                                                                                                                                              }
                                                                                                                                                              }
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "SSE stream of AI response with citations"
                                                                                             }
                                                                                 }
                                                               }
                                                  },
...
              },
...
}